impl: terminado modulo de music, search, autenticacion, home y user, terminada la implementacion de su V1

+++ src/main/java/com/essence/essencebackend/library/playlist/dto/PlaylistResponseDTO.java
@@ -15,6 +15,6 @@ public record PlaylistResponseDTO(
         Integer totalSongs,
 
         @JsonInclude(JsonInclude.Include.NON_NULL)
-        Integer totalLikes
+        Long totalLikes
 ) {
 }
